Update patch set 1
Patch Set 1:

The test files will probably not have the .md extension, but rather the format that you write your test cases in. In EPG, it is .cpp, .c or .ttcn. But if you want some real runnable test cases in this repo, something like Python could be a better choice, but it depends on what you are proficient in.

Of course, if we switch language, the formatting of the comment containing the trace information will have to be changed to be legal in that language.

Since the repo needs to include a runnable validation tool in order to demonstrate how T-Reqs would work, what do you think of my proposal that the example test cases and requirements in this repo should actually be test cases and requirements for that validation tool? 

If so, a hypothetical Python test case could look like (with the docstring written in the - for Python - common format REStructured Text):

(File: test_validation.py)

  def test_validate_tracing_simple():
  """Test that validation of tracing works in a simple case

  :feature #1

  :requirement REQ1
  :requirement REQ2
  """
      requirements = get_requirements('validation_requirements.md')
      test_cases = get_test_cases('test_validation.py')
      features = get_features_from_github()

      error_messages = validate_traces(requirements, test_cases, features)
      assert len(error_messages) == 0

Patch-set: 1
Reviewer: filip.lange@gmail.com <1014268@d5d70762-12d0-45a1-890d-524b12d3f735>