Make treqs into a python package
This makes test easier, since you can just do "import treqs" in your
test file. Otherwise you have to reference the exact directory and so
on, and it is much harder.

I changed the name of the scripts folder to "treqs", which is now the
name of the package.

This means that we can register the package on PyPi if we want to at a
later date, which makes it so that someone can just type "pip install
treqs" to install this application to their system.

Additionally, I added a command-line script "treqs", which makes it
possible to just type "treqs" anywhere to execute the validation.

setup.py requires you to specify author and license. I specified
Eric and an MIT license, feel free to suggest alternative values.

Change-Id: Ic0b485b783baae47e79988d343095db10562e403