Fix log file only created once per import treqs statement
Since the opening of the log file was in the class instead
of in the init method, when creating a second instance of
the class, it still tried to write to the same log file,
but since that log file had been closed it gave an error.

Also reshuffled the creation of the logs/ folder to avoid
having to do it in every class.

Change-Id: Ib3d9241a72650bedca8c04018c196d31c410dde0